#12864: Undo/Redo rounds column gap value
git-svn-id: svn://scribus.net/trunk/Scribus@19703 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pageitem.cpp
@@ -6700,9 +6700,9 @@ void PageItem::restorePasteText(SimpleState *ss, bool isUndo)
 void PageItem::restoreColumnsGap(SimpleState *ss, bool isUndo)
 {
 	if (isUndo)
-		ColGap = ss->getInt("OLD_COLUMNS");
+		ColGap = ss->getDouble("OLD_COLUMNS");
 	else
-		ColGap = ss->getInt("NEW_COLUMNS");
+		ColGap = ss->getDouble("NEW_COLUMNS");
 	update();
 }
